requirements.txt
pymongo
tensorflow
flask
Pillow
numpy